pyproject.toml
[project]
name = "sw-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
